Fix Stadia axis format: int16 sticks → uint8 to match real Stadia byte layout
macOS applies a fixed byte-offset mapping for known Stadia VID/PID 18d1:9400.
With int16 sticks (2 bytes each) the right-stick bytes landed at the offsets
macOS expects to hold trigger values, so the right stick was misread as triggers.

All six axes (LX, LY, RX, RY, R2/Z, L2/Rz) now use uint8 (0-255, 128=center),
matching the real Stadia Controller rev. A report format:
  offset 0-1: buttons, offset 2: hat, offsets 3-8: LX LY RX RY R2 L2

Changes:
- kDescHidGamepad: replaced 4×int16 + 2×uint8 with single 6×uint8 Input item
- GamepadReport: left_x/y, right_x/y changed from int16 to uint8
- ParseForGamepad: removed scaling lambda; copy raw bytes directly

+++ src/device_out.h
@@ -26,19 +26,19 @@ struct MouseReport {
 } __attribute__((packed));
 
 // Stadia Controller USB HID report.
-// Byte layout (matches kDescHidGamepad):
-//   buttons (uint16): bits 0-15, see ParseForGamepad() for Stadia assignment
+// Byte layout matches the real Stadia Controller rev. A (18d1:9400):
+//   buttons (uint16): 2 bytes — see ParseForGamepad() for bit assignment
 //   hat     (uint8):  bits 0-3 = hat (0=N…7=NW, 0xF=center), bits 4-7 = 0
-//   left_x, left_y, right_x, right_y (int16): sticks
-//   r2 (uint8): right trigger — Z,  SDL2 axis 4
-//   l2 (uint8): left trigger  — Rz, SDL2 axis 5
+//   left_x, left_y, right_x, right_y (uint8, 0-255, 128=center): sticks
+//   r2 (uint8, 0-255): right trigger — Z,  SDL2 axis 4
+//   l2 (uint8, 0-255): left  trigger — Rz, SDL2 axis 5
 struct GamepadReport {
   uint16_t buttons;
   uint8_t hat;
-  int16_t left_x;
-  int16_t left_y;
-  int16_t right_x;
-  int16_t right_y;
+  uint8_t left_x;
+  uint8_t left_y;
+  uint8_t right_x;
+  uint8_t right_y;
   uint8_t r2;   // Z  — right trigger, SDL2 axis 4
   uint8_t l2;   // Rz — left trigger,  SDL2 axis 5
 } __attribute__((packed));
